requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
django-allauth>=65.0,<66.0
django-axes>=7.0,<8.0
django-csp>=4.0,<5.0
django-encrypted-model-fields>=0.6,<1.0
django-environ>=0.11,<1.0
django-ratelimit>=4.1,<5.0
gunicorn>=23.0,<24.0
psycopg[binary]>=3.2,<4.0
celery>=5.4,<6.0
redis>=5.2,<6.0
sentry-sdk[django]>=2.18,<3.0
structlog>=24.4,<25.0
web3>=7.6,<8.0
argon2-cffi>=23.1,<24.0

# Dev / Test
pytest>=8.3,<9.0
pytest-django>=4.9,<5.0
pytest-cov>=6.0,<7.0
coverage>=7.6,<8.0
ruff>=0.8,<1.0
bandit>=1.8,<2.0
safety>=3.2,<4.0
factory-boy>=3.3,<4.0
